Cleanup
Split formatting out into it's own module

Clean up some variable names and pull out some anonymous functions.
